+++ src/pages/Mark/MarkList.jsx
@@ -1,46 +1,54 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
-import { useState, useEffect } from 'react';
-import Swal from 'sweetalert2';
-import Header from './../../components/Header';
-import Sidebar from './../../components/Sidebar';
-import Footer from './../../components/Footer';
+import React from "react";
+import { Link } from "react-router-dom";
+import { useState, useEffect } from "react";
+import Swal from "sweetalert2";
+import Header from "./../../components/Header";
+import Sidebar from "./../../components/Sidebar";
+import Footer from "./../../components/Footer";
 
 const MarkList = () => {
   const [markList, setMarkList] = useState([]);
+  const [userRole, setUserRole] = useState("");
   const [loading, setLoading] = useState(true);
   const [currentPage, setCurrentPage] = useState(1);
   const dataPerPage = 5;
 
+  useEffect(() => {
+    let role = localStorage.getItem("role");
+    setUserRole(role);
+  });
+
   const deleteMark = (e, id) => {
     e.preventDefault();
     const Clicked = e.currentTarget;
-    Clicked.innerText = 'deleting';
+    Clicked.innerText = "deleting";
 
-    fetch(`http://127.0.0.1:8000/api/marks/${id}`, {
-      headers: {
-        Accept: 'application/json',
-      },
-      method: 'DELETE',
-    })
-      .then((response) => response.json())
-      .then((response) => {
-        console.info(response);
-        Swal.fire('Success', response?.message, 'success');
-        Clicked.closest('tr').remove();
+    if (confirm(`Are you sure you want to delete mark id ${id}?`)) {
+      fetch(`http://127.0.0.1:8000/api/marks/${id}`, {
+        headers: {
+          Accept: "application/json",
+        },
+        method: "DELETE",
       })
-      .catch((error) => {
-        console.error(error);
-        Swal.fire('Warning', response?.message, 'warning');
-      });
+        .then((response) => response.json())
+        .then((response) => {
+          console.info(response);
+          Swal.fire("Success", response?.message, "success");
+          Clicked.closest("tr").remove();
+        })
+        .catch((error) => {
+          console.error(error);
+          Swal.fire("Warning", response?.message, "warning");
+        });
+    }
   };
 
   useEffect(() => {
-    fetch('http://127.0.0.1:8000/api/marks?', {
+    fetch("http://127.0.0.1:8000/api/marks?", {
       headers: {
-        Accept: 'application/json',
+        Accept: "application/json",
       },
-      method: 'GET',
+      method: "GET",
     })
       .then((response) => response.json())
       .then((response) => {
@@ -60,146 +68,268 @@ const MarkList = () => {
   const npage = Math.ceil(markList?.length / dataPerPage);
   const numbers = [...Array(npage + 1).keys()].slice(1);
 
-  return (
-    <>
-      <div>
-        <Header />
-      </div>
-      <div className="d-flex">
-        <div className="w-auto position-sticky">
-          <Sidebar />
+  if (userRole === "1") {
+    return (
+      <>
+        <div>
+          <Header />
         </div>
-        <div className="col overflow-hidden">
-          <div className="container px-4">
-            <div className="card">
-              <div className="card-header">
-                <h4>Mark List</h4>
-                <Link
-                  to="/dashboard/marks/create"
-                  className="btn btn-primary btn-sm float-end"
-                >
-                  Add Mark
-                </Link>
-              </div>
-              <div className="page-system mt-4">
-                <nav>
-                  <ul className="pagination">
-                    <li className="page-item">
-                      <a href="#" className="page-link" onClick={prePage}>
-                        Prev
-                      </a>
-                    </li>
-                    {numbers.map((n, i) => {
-                      return (
-                        <li
-                          className={`page-item ${
-                            currentPage === n ? 'active' : ''
-                          }`}
-                          key={i}
-                        >
-                          <a
-                            href="#"
-                            className="page-link"
-                            onClick={() => changeCurrentPage(n)}
+        <div className="d-flex">
+          <div className="w-auto position-sticky">
+            <Sidebar />
+          </div>
+          <div className="col overflow-hidden">
+            <div className="container px-4">
+              <div className="card">
+                <div className="card-header">
+                  <h4>Mark List</h4>
+                  <Link
+                    to="/dashboard/marks/create"
+                    className="btn btn-primary btn-sm float-end"
+                  >
+                    Add Mark
+                  </Link>
+                </div>
+                <div className="page-system mt-4">
+                  <nav>
+                    <ul className="pagination">
+                      <li className="page-item">
+                        <a href="#" className="page-link" onClick={prePage}>
+                          Prev
+                        </a>
+                      </li>
+                      {numbers.map((n, i) => {
+                        return (
+                          <li
+                            className={`page-item ${
+                              currentPage === n ? "active" : ""
+                            }`}
+                            key={i}
                           >
-                            {n}
-                          </a>
-                        </li>
-                      );
-                    })}
-                    <li className="page-item">
-                      <a href="#" className="page-link" onClick={nextPage}>
-                        Next
-                      </a>
-                    </li>
-                  </ul>
-                </nav>
-              </div>
-              <div className="card-body">
-                <table className="table table-striped">
-                  <thead>
-                    <tr>
-                      <th scope="col">ID</th>
-                      <th scope="col">Mark</th>
-                      <th scope="col">Grade Point</th>
-                      <th scope="col">Class Name</th>
-                      <th scope="col">Student Name</th>
-                      <th scope="col">Exam</th>
-                      <th scope="col">Section Name</th>
-                      <th scope="col">Subject Name</th>
-                      <th scope="col">Comment</th>
-                      <th scope="col">Show</th>
-                      <th scope="col">Edit</th>
-                      <th scope="col">Delete</th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                    {records?.map((mark) => {
-                      return (
-                        <tr key={mark?.id}>
-                          <td>{mark?.id}</td>
-                          <td>{mark?.marks}</td>
-                          <td>{mark?.grade_point}</td>
-                          <td>{mark?.class?.name}</td>
-                          <td>{mark?.user?.name}</td>
-                          <td>{mark?.exam?.name}</td>
-                          <td>{mark?.section?.name}</td>
-                          <td>{mark?.subject?.name}</td>
-                          <td>{mark?.comment}</td>
-                          <td>
-                            <Link
-                              to={`/dashboard/marks/${mark?.id}/show`}
-                              className="btn btn-primary btn-sm"
+                            <a
+                              href="#"
+                              className="page-link"
+                              onClick={() => changeCurrentPage(n)}
                             >
-                              Show
-                            </Link>
-                          </td>
-                          <td>
-                            <Link
-                              to={`/dashboard/marks/${mark?.id}/edit`}
-                              className="btn btn-success btn-sm"
+                              {n}
+                            </a>
+                          </li>
+                        );
+                      })}
+                      <li className="page-item">
+                        <a href="#" className="page-link" onClick={nextPage}>
+                          Next
+                        </a>
+                      </li>
+                    </ul>
+                  </nav>
+                </div>
+                <div className="card-body">
+                  <table className="table table-striped">
+                    <thead>
+                      <tr>
+                        <th scope="col">ID</th>
+                        <th scope="col">Mark</th>
+                        <th scope="col">Grade Point</th>
+                        <th scope="col">Class Name</th>
+                        <th scope="col">Student Name</th>
+                        <th scope="col">Exam</th>
+                        <th scope="col">Section Name</th>
+                        <th scope="col">Subject Name</th>
+                        <th scope="col">Comment</th>
+                        <th scope="col">Show</th>
+                        <th scope="col">Edit</th>
+                        <th scope="col">Delete</th>
+                      </tr>
+                    </thead>
+                    <tbody>
+                      {records?.map((mark) => {
+                        return (
+                          <tr key={mark?.id}>
+                            <td>{mark?.id}</td>
+                            <td>{mark?.marks}</td>
+                            <td>{mark?.grade_point}</td>
+                            <td>{mark?.class?.name}</td>
+                            <td>{mark?.user?.name}</td>
+                            <td>{mark?.exam?.name}</td>
+                            <td>{mark?.section?.name}</td>
+                            <td>{mark?.subject?.name}</td>
+                            <td>{mark?.comment}</td>
+                            <td>
+                              <Link
+                                to={`/dashboard/marks/${mark?.id}/show`}
+                                className="btn btn-primary btn-sm"
+                              >
+                                Show
+                              </Link>
+                            </td>
+                            <td>
+                              <Link
+                                to={`/dashboard/marks/${mark?.id}/edit`}
+                                className="btn btn-success btn-sm"
+                              >
+                                Edit
+                              </Link>
+                            </td>
+                            <td
+                              type="button"
+                              onClick={(e) => deleteMark(e, mark?.id)}
+                              className="btn btn-danger btn-sm"
                             >
-                              Edit
-                            </Link>
-                          </td>
-                          <td
-                            type="button"
-                            onClick={(e) => deleteMark(e, mark?.id)}
-                            className="btn btn-danger btn-sm"
-                          >
-                            Delete
-                          </td>
-                        </tr>
-                      );
-                    })}
-                  </tbody>
-                </table>
+                              Delete
+                            </td>
+                          </tr>
+                        );
+                      })}
+                    </tbody>
+                  </table>
+                </div>
               </div>
             </div>
           </div>
         </div>
-      </div>
-      <div>
-        <Footer />
-      </div>
-    </>
-  );
+        <div>
+          <Footer />
+        </div>
+      </>
+    );
 
-  function prePage(e) {
-    e.preventDefault();
-    if (currentPage !== 1) {
-      setCurrentPage(currentPage - 1);
-    }
-  }
-
-  function changeCurrentPage(id) {
-    setCurrentPage(id);
-  }
-
-  function nextPage(e) {
-    if (currentPage !== npage) {
+    function prePage(e) {
       e.preventDefault();
-      setCurrentPage(currentPage + 1);
+      if (currentPage !== 1) {
+        setCurrentPage(currentPage - 1);
+      }
+    }
+
+    function changeCurrentPage(id) {
+      setCurrentPage(id);
+    }
+
+    function nextPage(e) {
+      if (currentPage !== npage) {
+        e.preventDefault();
+        setCurrentPage(currentPage + 1);
+      }
+    }
+  } else if(userRole === '2' || userRole === '3' || userRole === '4') {
+    return (
+      <>
+        <div>
+          <Header />
+        </div>
+        <div className="d-flex">
+          <div className="w-auto position-sticky">
+            <Sidebar />
+          </div>
+          <div className="col overflow-hidden">
+            <div className="container px-4">
+              <div className="card">
+                <div className="card-header">
+                  <h4>Mark List</h4>
+                </div>
+                <div className="page-system mt-4">
+                  <nav>
+                    <ul className="pagination">
+                      <li className="page-item">
+                        <a href="#" className="page-link" onClick={prePage}>
+                          Prev
+                        </a>
+                      </li>
+                      {numbers.map((n, i) => {
+                        return (
+                          <li
+                            className={`page-item ${
+                              currentPage === n ? "active" : ""
+                            }`}
+                            key={i}
+                          >
+                            <a
+                              href="#"
+                              className="page-link"
+                              onClick={() => changeCurrentPage(n)}
+                            >
+                              {n}
+                            </a>
+                          </li>
+                        );
+                      })}
+                      <li className="page-item">
+                        <a href="#" className="page-link" onClick={nextPage}>
+                          Next
+                        </a>
+                      </li>
+                    </ul>
+                  </nav>
+                </div>
+                <div className="card-body">
+                  <table className="table table-striped">
+                    <thead>
+                      <tr>
+                        <th scope="col">ID</th>
+                        <th scope="col">Mark</th>
+                        <th scope="col">Grade Point</th>
+                        <th scope="col">Class Name</th>
+                        <th scope="col">Student Name</th>
+                        <th scope="col">Exam</th>
+                        <th scope="col">Section Name</th>
+                        <th scope="col">Subject Name</th>
+                        <th scope="col">Comment</th>
+                        <th scope="col">Show</th>
+                      </tr>
+                    </thead>
+                    <tbody>
+                      {records?.map((mark) => {
+                        return (
+                          <tr key={mark?.id}>
+                            <td>{mark?.id}</td>
+                            <td>{mark?.marks}</td>
+                            <td>{mark?.grade_point}</td>
+                            <td>{mark?.class?.name}</td>
+                            <td>{mark?.user?.name}</td>
+                            <td>{mark?.exam?.name}</td>
+                            <td>{mark?.section?.name}</td>
+                            <td>{mark?.subject?.name}</td>
+                            <td>{mark?.comment}</td>
+                            <td>
+                              <Link
+                                to={`/dashboard/marks/${mark?.id}/show`}
+                                className="btn btn-primary btn-sm"
+                              >
+                                Show
+                              </Link>
+                            </td>
+                          </tr>
+                        );
+                      })}
+                    </tbody>
+                  </table>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+        <div>
+          <Footer />
+        </div>
+      </>
+    );
+
+    function prePage(e) {
+      e.preventDefault();
+      if (currentPage !== 1) {
+        setCurrentPage(currentPage - 1);
+      }
+    }
+
+    function changeCurrentPage(id) {
+      setCurrentPage(id);
+    }
+
+    function nextPage(e) {
+      if (currentPage !== npage) {
+        e.preventDefault();
+        setCurrentPage(currentPage + 1);
+      }
     }
   }
 };
